+++ src/pages/shop/collect.tsx
@@ -1,29 +1,37 @@
+import { useAppSelector } from 'app/hook';
 import DropDownFilter from 'components/common/DropDownFilter';
+import EmptyContent from 'components/common/EmptyContent';
 import ShopCard from 'components/common/ShopCard';
-import { shopApi, useGetShopByBookmarkQuery } from 'features/shops/shopApi';
-import { useEffect, useState } from 'react';
+import { useGetShopByBookmarkQuery } from 'features/shops/shopApi';
+import { selectIsLogin } from 'features/users/userSlice';
+import { useState } from 'react';
 import styled from 'styled-components';
 import { applyMediaQuery } from 'styles/mediaQuery';
 import { theme } from 'styles/theme';
 import { ShopBookMarkRequestType } from 'types/shop';
 
+const emptyContentData = {
+  title: '저장한 소품샵',
+  src: '/assets/img_shopNoContent.png',
+  label: '아직 저장한 소품샵이 없어요',
+  subLabel: '취향저격 소품샵 찾으러 가볼까요?',
+  button: '테마별 소품샵 보러가기',
+  buttonUrl: '/shop/theme/아기자기한',
+};
+
 function Collect() {
+  const isLogin = useAppSelector(selectIsLogin);
   const [currentSort, setCurrentSort] = useState<ShopBookMarkRequestType>('save');
-  const { data: collectShopList } = useGetShopByBookmarkQuery({
-    sort: currentSort,
-    offset: 1,
-    limit: 12,
-  });
-
-  const [fetchCollectShops] = shopApi.useLazyGetShopByBookmarkQuery();
-  const [currentList, setCurrentList] = useState([]);
-
-  useEffect(() => {
-    (async function () {
-      const data = await fetchCollectShops({ sort: currentSort, offset: 1, limit: 12 });
-      console.log(data);
-    })();
-  }, [currentSort]);
+  const { data: collectShopList } = useGetShopByBookmarkQuery(
+    {
+      sort: currentSort,
+      offset: 1,
+      limit: 12,
+    },
+    {
+      skip: !isLogin,
+    },
+  );
 
   const filterProps = [
     {
@@ -42,14 +50,20 @@ function Collect() {
 
   return (
     <StyledContainer>
-      <h2>저장한 소품샵</h2>
-      <StyledFilterWrapper>
-        <DropDownFilter pageType="collect" filterProps={filterProps} />
-      </StyledFilterWrapper>
-      <StyledCardWrapper>
-        {collectShopList &&
-          collectShopList.map((shop) => <ShopCard key={shop.shopId} cardData={shop} />)}
-      </StyledCardWrapper>
+      {collectShopList?.length ? (
+        <>
+          <h2>저장한 소품샵</h2>
+          <StyledFilterWrapper>
+            <DropDownFilter pageType="collect" filterProps={filterProps} />
+          </StyledFilterWrapper>
+          <StyledCardWrapper>
+            {collectShopList &&
+              collectShopList.map((shop) => <ShopCard key={shop.shopId} cardData={shop} />)}
+          </StyledCardWrapper>
+        </>
+      ) : (
+        <EmptyContent emptyContentData={emptyContentData} />
+      )}
     </StyledContainer>
   );
 }
